requirements.txt
# Основные библиотеки
numpy>=1.21.0,<2.0.0
pandas>=1.3.0,<2.0.0
scikit-learn>=1.0.0,<2.0.0
scipy>=1.7.0,<2.0.0

# Обработка текста
spacy>=3.4.0,<4.0.0
pymorphy2>=0.9.1
pymorphy2-dicts-ru
nltk>=3.6.0,<4.0.0
regex>=2021.0.0

# Векторизация
gensim>=4.2.0,<5.0.0
rank-bm25>=0.2.2

# Кластеризация
hdbscan>=0.8.29

# Визуализация
matplotlib>=3.5.0,<4.0.0
seaborn>=0.11.0,<0.13.0
plotly>=5.0.0,<6.0.0
umap-learn>=0.5.3

# Веб-интерфейс
streamlit>=1.22.0,<2.0.0

# BPE обработка
subword-nmt>=0.3.8

# Дополнительные
jupyter>=1.0.0,<2.0.0
tqdm>=4.62.0,<5.0.0
threadpoolctl>=3.1.0